# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AccountResponse:
    properties:
      account_number:
        type: string
      account_type:
        $ref: '#/definitions/enums.AccountType'
      balance:
        type: number
      client_id:
        type: string
      client_name:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/enums.AccountStatus'
      updated_at:
        type: string
    type: object
  dto.AlertResponse:
    properties:
      ai_explanation:
        type: string
      alert_type:
        $ref: '#/definitions/enums.AlertType'
      blockchain_tx:
        type: string
      client_id:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      severity:
        $ref: '#/definitions/enums.AlertSeverity'
      status:
        $ref: '#/definitions/enums.AlertStatus'
      transaction_code:
        type: string
      transaction_id:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      token:
        type: string
    type: object
...
  dto.CreateAccountRequest:
    properties:
      account_number:
        type: string
      account_type:
        $ref: '#/definitions/enums.AccountType'
      client_id:
        type: string
      currency:
        type: string
    required:
    - account_number
    - account_type
    - client_id
    - currency
    type: object
  dto.CreateAlertRequest:
    properties:
      ai_explanation:
        type: string
      alert_type:
        $ref: '#/definitions/enums.AlertType'
      client_id:
        type: string
      description:
        type: string
      severity:
        $ref: '#/definitions/enums.AlertSeverity'
      transaction_id:
        type: string
    required:
    - alert_type
    - client_id
    - description
    - severity
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  dto.UpdateAccountRequest:
    properties:
      status:
        $ref: '#/definitions/enums.AccountStatus'
    required:
    - status
    type: object
  dto.UpdateAlertStatusRequest:
    properties:
      reviewed_by:
        type: string
      status:
        $ref: '#/definitions/enums.AlertStatus'
    required:
    - reviewed_by
    - status
    type: object
...
  enums.AccountStatus:
    enum:
    - active
    - blocked
    - closed
    type: string
    x-enum-varnames:
    - AccountStatusActive
    - AccountStatusBlocked
    - AccountStatusClosed
  enums.AccountType:
    enum:
    - checking
    - savings
    - credit_card
    type: string
    x-enum-varnames:
    - AccountTypeChecking
    - AccountTypeSavings
    - AccountTypeCreditCard
  enums.AlertSeverity:
    enum:
    - low
    - medium
    - high
    - critical
    type: string
    x-enum-varnames:
    - AlertSeverityLow
    - AlertSeverityMedium
    - AlertSeverityHigh
    - AlertSeverityCritical
  enums.AlertStatus:
    enum:
    - pending
    - reviewed
    - dismissed
    - escalated
    type: string
    x-enum-varnames:
    - AlertStatusPending
    - AlertStatusReviewed
    - AlertStatusDismissed
    - AlertStatusEscalated
  enums.AlertType:
    enum:
    - unusual_amount
    - unusual_location
    - high_frequency
    - fraud_suspicion
    type: string
    x-enum-varnames:
    - AlertTypeUnusualAmount
    - AlertTypeUnusualLocation
    - AlertTypeHighFrequency
    - AlertTypeFraudSuspicion
...
  response.AccountListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AccountResponse'
        type: array
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
...
  response.AccountResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/dto.AccountResponse'
    type: object
  response.AlertListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AlertResponse'
        type: array
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
  response.AlertResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/dto.AlertResponse'
    type: object
...
  response.ErrorDetails:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetails'
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Fraud detection and transaction monitoring API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Zentinel API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      description: List accounts with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AccountListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account for a client
      parameters:
      - description: Create Account Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.AccountResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
  /api/v1/accounts/{id}:
    get:
      description: Get account details by ID or Account Number
      parameters:
      - description: Account ID or Number
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AccountResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account by ID or Number
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update account details (e.g. status)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Account Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AccountResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - accounts
  /api/v1/alerts:
    get:
      description: List alerts with filtering and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Severity
        in: query
        name: severity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AlertListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List alerts
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Create a new alert manually
      parameters:
      - description: Create Alert Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.AlertResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new alert
      tags:
      - alerts
  /api/v1/alerts/{id}:
    get:
      description: Get alert details by ID
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AlertResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get alert by ID
      tags:
      - alerts
  /api/v1/alerts/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Alert Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AlertResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update alert status
      tags:
      - alerts
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"